Do not dispose server certificate
Fixes issue #74
The sslProtocol is cloned for every new client connection, this clone is
made by calling the constructor with the fields in the current object. This
means that the Certificate object is shared between all clones.

So when a client disconnects, we shouldn't dispose the certificate.
Because this means that all future connections will fail because thet
can't use the certificate.