Remove call to socket.Poll in the isConnected func
Because the if statement was broken poll was always called, even if the
poll parameter was set to false. This caused problems when calling the
isConnected function in a loop.

The poll function increased accuracy in certain situations, but with
keep alive support and all protocols auto disconnecting when receive
returns an empty byte array this function is mostly unused.

If a user still wants to use the poll functionality, then this is still
possible by accessing the socket directly. (client.BaseSocket)